Stat collection

+++ src/iblock/bitcoin/apps/BlockchainManager.cc
@@ -18,8 +18,18 @@ void BlockchainManager::initialize(int stage)
 	case 0:
 		AppBase::initialize(0);
 
+		networkDifficultySignal = registerSignal("networkDifficulty");
+		mainBranchAppendSignal = registerSignal("mainBranchAppend");
+		sideBranchAppendSignal = registerSignal("sideBranchAppend");
+		mainBranchSwapSignal = registerSignal("mainBranchSwap");
+		newSideBranchSignal = registerSignal("newSideBranch");
+		orphanCountSignal = registerSignal("orphanCount");
+		mainBranchHeightSignal = registerSignal("mainBranchHeight");
+		mainBranchTransactionCountSignal = registerSignal("mainBranchTransactionCount");
+		sideBranchLengthSignal = registerSignal("sideBranchLength");
+		duplicateBlockSignal = registerSignal("duplicateBlock");
+
 		nodeManager = check_and_cast<NodeManager* >(getModuleByPath(par("nodeManagerModule").stringValue()));
-		gbm = check_and_cast<GBM*>(getModuleByPath(par("gbmModule").stringValue()));
 
 		chainHistory = par("chainHistory").intValue();
 
@@ -88,15 +98,42 @@ uint32_t BlockchainManager::getCurrentHeight() const
 	return this->getCurrentBlock()->getHeight();
 }
 
+double BlockchainManager::computeNewDifficulty() const
+{
+	double oldDifficulty = getCurrentDifficulty();
+	simtime_t actualTime = getCurrentBlockHeader()->getTime() - adjustmentBlockTime;
+	simtime_t expectedTime = SimTime(2016 * 10 * 60, SimTimeUnit::SIMTIME_S);
+	double ratio = actualTime / expectedTime;
+	if (ratio < 0.25)
+		ratio = 0.25;
+	else if (ratio > 4.0)
+		ratio = 4.0;
+	return oldDifficulty / ratio;
+}
+
+double BlockchainManager::getCurrentDifficulty() const
+{
+	Hash highestTarget = Hash(0x1d00ffff);
+	Hash oldTarget = getCurrentTargetNBits();
+	unsigned int expDiff = (highestTarget.e() - oldTarget.e()) << 3;
+	return std::ldexp(highestTarget.m() / static_cast<double>(oldTarget.m()), expDiff);
+}
+
 Hash BlockchainManager::getNextTargetNBits() const
 {
-	return this->getCurrentTargetNBits(); // TODO
+	if (getCurrentHeight() % 2016 != 2015)
+		return getCurrentTargetNBits();
+	double newDifficulty = computeNewDifficulty();
+	double newTarget = std::ldexp(0xffff, 8 * (0x1d - 3)) / newDifficulty;
+	return Hash::fromBigDouble(newTarget);
 }
 
 std::shared_ptr<const Block> BlockchainManager::findForkBlock(std::shared_ptr<const Block> a, std::shared_ptr<const Block> b) const
 {
 	if (a == b)
 		return a;
+	if (!a || !b)
+		return nullptr;
 	if (a->getHeight() > b->getHeight())
 		return findForkBlock(a->getPrevBlock(), b);
 	if (a->getHeight() < b->getHeight())
@@ -117,12 +154,13 @@ void BlockchainManager::unconfirmWalletUtxos(std::shared_ptr<const Block> block)
 
 void BlockchainManager::confirmWalletUtxos(std::shared_ptr<const Block> block) const
 {
+	bool genesis = block->getHeight() == 0;
 	for (Wallet* wallet : wallets) {
 		const std::unordered_set<std::shared_ptr<const TransactionOutput>>* utxos = block->getUtxos(wallet);
 		if (!utxos)
 			continue;
 		for (std::shared_ptr<const TransactionOutput> utxo : *utxos)
-			wallet->confirmUtxo(utxo);
+			wallet->confirmUtxo(utxo, genesis);
 	}
 }
 
@@ -134,10 +172,14 @@ void BlockchainManager::cleanup(uint32_t history)
 
 	for (auto it = branches.begin(); it != branches.end();) {
 		std::shared_ptr<const Block> b = *it;
-		if (b->getHeight() < height - history)
+		if (b->getHeight() < height - history) {
+			std::shared_ptr<const Block> forkBlock = findForkBlock(b, mainBranch);
+			if (forkBlock)
+				emit(sideBranchLengthSignal, b->getHeight() - forkBlock->getHeight());
 			it = branches.erase(it);
-		else
+		} else {
 			++it;
+		}
 	}
 
 	for (auto it = orphans.begin(); it != orphans.end();)
@@ -149,6 +191,14 @@ void BlockchainManager::cleanup(uint32_t history)
 	std::shared_ptr<const Block> block = mainBranch;
 	for (uint32_t curHeight = height; curHeight >= height - history && block; curHeight--)
 		block = block->getPrevBlock();
+	if (!block)
+		return;
+	for (auto it = branches.begin(); it != branches.end(); ++it) {
+		std::shared_ptr<const Block> b = *it;
+		if (b == mainBranch)
+			continue;
+		block = findForkBlock(b, block);
+	}
 	if (block)
 		const_cast<BlockHeader*>(block->getHeader())->deletePrevBlock();
 }
@@ -171,6 +221,10 @@ void BlockchainManager::appendBlock(std::shared_ptr<const Block> block)
 			type |= 0x02;
 		switch (type) {
 		[[likely]] case 0x01: // block further extends the main branch
+			if (height % 2016 == 0) {
+				adjustmentBlockTime = block->getHeader()->getTime();
+				emit(networkDifficultySignal, computeNewDifficulty());
+			}
 			mainBranch = block;
 			mempoolManager->removeBlockTransactions(mainBranch);
 			confirmWalletUtxos(mainBranch);
@@ -193,6 +247,10 @@ void BlockchainManager::appendBlock(std::shared_ptr<const Block> block)
 				toAdd.push_back(b);
 			while (!toAdd.empty()) {
 				std::shared_ptr<const Block> b = toAdd.back();
+				if (height % 2016 == 0) {
+					adjustmentBlockTime = block->getHeader()->getTime();
+					emit(networkDifficultySignal, computeNewDifficulty());
+				}
 				mainBranch = b;
 				mempoolManager->removeBlockTransactions(b);
 				confirmWalletUtxos(b);
@@ -218,7 +276,7 @@ void BlockchainManager::appendBlock(std::shared_ptr<const Block> block)
 		throw cRuntimeError("Genesis block already exists");
 
 	for (const auto& branch : branches)
-		for (std::shared_ptr<const Block> cur = branch; cur->getHeight() >= height; cur = cur->getPrevBlock()) {
+		for (std::shared_ptr<const Block> cur = branch; cur && cur->getHeight() >= height; cur = cur->getPrevBlock()) {
 			if (cur == block) {
 				onDuplicateBlock(block);
 				return; // duplicate block
@@ -250,11 +308,56 @@ void BlockchainManager::addBlock(std::shared_ptr<Block> block)
 {
 	Enter_Method("addBlock()");
 
-	onBeforeMinedBlockAppend(block);
+	this->onBeforeMinedBlockAppend(block);
 
 	appendBlock(std::const_pointer_cast<const Block>(block));
 
 	onAfterMinedBlockAppend(block);
+}
+
+void BlockchainManager::addGenesisBlock(std::shared_ptr<const Block> block)
+{
+	Enter_Method_Silent("addGenesisBlock()");
+	if (mainBranch)
+		throw cRuntimeError("Genesis block already set");
+	adjustmentBlockTime = block->getHeader()->getTime();
+	mainBranch = block;
+	branches.insert(mainBranch);
+	emit(mainBranchHeightSignal, 0U);
+	emit(mainBranchTransactionCountSignal, static_cast<int>(mainBranch->getTxnCount()));
+	emit(networkDifficultySignal, getCurrentDifficulty());
+	confirmWalletUtxos(block);
+}
+
+void BlockchainManager::onNewMainBranch(std::shared_ptr<const Block> oldBranch, std::shared_ptr<const Block> forkBlock)
+{
+	emit(mainBranchSwapSignal, 1U);
+	emit(mainBranchHeightSignal, mainBranch->getHeight());
+	for (std::shared_ptr<const Block> b = oldBranch; b != forkBlock; b = b->getPrevBlock())
+		emit(mainBranchTransactionCountSignal, -static_cast<int>(b->getTxnCount()));
+	for (std::shared_ptr<const Block> b = mainBranch; b != forkBlock; b = b->getPrevBlock())
+		emit(mainBranchTransactionCountSignal, static_cast<int>(b->getTxnCount()));
+	cleanup(chainHistory);
+}
+
+void BlockchainManager::onMainBranchAppend()
+{
+	emit(mainBranchAppendSignal, 1U);
+	emit(mainBranchHeightSignal, mainBranch->getHeight());
+	emit(mainBranchTransactionCountSignal, static_cast<int>(mainBranch->getTxnCount()));
+	cleanup(chainHistory);
+}
+
+void BlockchainManager::finish()
+{
+	for (auto it = branches.begin(); it != branches.end(); ++it) {
+		std::shared_ptr<const Block> branch = *it;
+		if (branch == mainBranch)
+			continue;
+		std::shared_ptr<const Block> forkBlock = findForkBlock(branch, mainBranch);
+		if (forkBlock)
+			emit(sideBranchLengthSignal, branch->getHeight() - forkBlock->getHeight());
+	}
 }
 
 }
